Fix types

+++ src/random_pop.rs
@@ -30,35 +30,35 @@ pub fn random_population<P: RandNode+Clone+Sync, F: Fitness+Sized+Send, R: Rng>(
     let mut ret = Population::new(n, 0);
     for i in 0..n {
         let height = 1 + i / (n / RANDOMPOP_MAX_HEIGHT);
-        ret.add(P::rand(TargetHeight::fixed(height as u32), rng));
+        ret.add(P::rand(TargetHeight::fixed(height as i32), rng));
     }
     ret
 }
 
 #[derive(Copy,Clone)]
 pub struct TargetHeight {
-    current_level: u32,
-    per_level: u32
+    current_level: i32,
+    per_level: i32
 }
 
 impl TargetHeight {
-    pub fn fixed(target_height: u32) -> TargetHeight {
+    pub fn fixed(target_height: i32) -> TargetHeight {
         TargetHeight {
             current_level: 0,
             per_level: 100 / max(target_height, 1)
         }
     }
 
-    pub fn randomized(max_height: u32, rng: &mut Rng) -> TargetHeight {
-        TargetHeight::fixed(1 + rng.next_u32() % (max_height - 1))
+    pub fn randomized(max_height: i32, rng: &mut Rng) -> TargetHeight {
+        TargetHeight::fixed(1 + rng.next_u32() as i32 % (max_height - 1))
     }
 
     pub fn internal(&self) -> u32 {
-        max(100 - self.per_level * self.current_level, MIN_WEIGHT)
+        max(100 - self.per_level * self.current_level, MIN_WEIGHT) as u32
     }
 
     pub fn leaf(&self) -> u32 {
-        max(self.per_level * self.current_level, MIN_WEIGHT)
+        max(self.per_level * self.current_level, MIN_WEIGHT) as u32
     }
 
     pub fn next_level(&self) -> TargetHeight {
@@ -74,4 +74,4 @@ impl TargetHeight {
 ///
 /// We need at least some weight, in case we need to make a leaf node
 /// but only internal nodes are available at that point in the tree.
-const MIN_WEIGHT : u32 = 1;
+const MIN_WEIGHT : i32 = 1;
